requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
colorama
PyYAML
langchain-openai
langchain-core
langchain-community
langchain-qdrant
langchain-google-genai
langchain-huggingface
langchain-ollama
qdrant-client
openai
google-generativeai
nltk
python-docx
PyPDF2